Load and Evaluate the TF lite model
develop a function to load the TF-lite model by creating an instance of the interpreter, allocating the tensors for inputs and outputs, invoking the interpreter, and evaluating 
the output results using MAE as a metric.   
define a function to get the model size and check if all constraints are met ( model size, temp_MAE, hum_MAE) for each version. 

# Note: this function can be ( should be ) executed on the edge device (Raspberrypi in our case)  but since we are not measuring latency the accuracy should not be affected