Building NN models
in this part of the script, we created 2 models, the first one is based on CNN layers, the second one is based on depth-wise CNN layers. We saved the model to a dictionary to call them in the next part of the code by accessing the keys of the dictionary. 
we also implemented 'Structural Pruning' to the layers controlled by alpha (the width multiplier), which is an input argument to the application.